requirements.txt
Flask==3.1.1
python-whois==0.9.5